==> 0_clean_species_trait_sheets.Rmd <==
---
title: "Clean species trait workbook"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)
### helper functions and setup
library(oharac) ### remotes::install_github('oharac/oharac')
library(tidyverse)
library(here)

```

# Summary

Read in Excel spreadsheets of taxonomic traits filled in by taxon experts.  Process into a long/tidy format with checks on validity of category values.  Identify categories that were assigned non-valid values.

# Data

`_raw_data/all_taxa_trait_data.xlsx' is the raw spreadsheet prepared by Nathalie Butt from the various submissions of the taxa-group experts.  Basic format:

Columns `category`, `sub-category`, and `category units` are standard.  Additional columns correspond to species or species groups, with perhaps additional idiosyncratic columns populated with totals, notes, comments, etc.

# Methods

## Set up cleaner functions

to aid with the processing of messy data

```{r}
clean_up_mess <- function(df) {
  df %>%
    mutate(trait_val_clean = str_trim(trait_val_clean)) %>%
    filter(trait_val_clean != '') %>%
    select(category, trait, trait_value, trait_val_clean, spp_gp, sheet) %>%
    distinct()
}

clean_trait <- function(x, drop_parens = TRUE) {
  ### replace egg-layer (unattended) to drop parens:
  y <- str_replace(x, '\\(unattended\\)', 'unattended')

  ### drop parens if needed
  if(drop_parens) {
    y <- y %>%
      str_replace_all('\\((.+?)\\)', '')
  }  
  ### then drop internal numeric commas and proceed from there
  y <- str_replace_all(y, '(?<=[0-9]),(?=[0-9])', '') %>%
    ### then: drop all non-alphanumeric and a few key punctuation:
    str_replace_all('[^0-9a-zA-Z<>,;\\-\\.\\(\\)/ ]', '') %>% 
    ### lower case; do it after dropping any weird non-ascii characters:
    tolower() %>% 
    str_trim() %>%
    str_replace_all(' +', ' ') %>% ### drop multi spaces
    str_replace_all('n/a', 'na') %>%
    ### convert remaining commas and slashes to semicolons:
    str_replace_all('[,/]', ';') %>%
    ### drop spaces after numbers e.g. 3 mm -> 3mm:
    str_replace_all('(?<=[0-9]) ', '') %>%
    ### drop spaces before or after punctuation (non-alphanumeric):
    str_replace_all(' (?=[^a-z0-9\\(])|(?<=[^a-z0-9\\)]) ', '') %>%
    ### manually fix some valid slashes:
    str_replace_all('nearly sessile;sedentary', 'nearly sessile/sedentary') %>%
    str_replace_all('live birth;egg care', 'live birth/egg care') %>%
    str_replace_all('chitin;caco3mix', 'chitin/caco3 mix')
    
  return(y)
}
```

## Process Nat's key to create a final key.  

Add an "unknown" value for all traits, to catch explicitly stated unknown values.

```{r}
traitvals_df_raw <- read_csv(here('_raw_data/trait_val_key.csv')) %>%
  select(category, trait, trait_value = trait_vals) %>%
  filter(!(is.na(category) & is.na(trait))) %>%
  mutate(category = tolower(category), 
         trait = tolower(trait),
         trait_value = clean_trait(trait_value))
  
traitvals_df_prehabs <- traitvals_df_raw %>%
  mutate(trait_value = paste0(trait_value, ';unknown'),
         trait_value = str_split(trait_value, pattern = ';')) %>%
  unnest(trait_value) %>%
  ### trim whitespace, replace the temporary slash fixes
  mutate(trait_value = str_trim(trait_value)) %>%
           ### undo the temp slash fix
  filter(!is.na(trait_value)) %>%
  filter(trait_value != '')

### Read in the first column of habitats from the list.  Paste together with
### semicolon separator.  Set to lower case.
habs <- readxl::read_excel(here('_raw_data/xlsx/hab_list.xlsx')) %>%
  janitor::clean_names() %>%
  .$hab_type %>%
  tolower() %>%
  paste(collapse = ';')
    
### Separate out habitat rows, attach valid habitats, bind back to original.
traitvals_hab <- traitvals_df_prehabs %>%
  filter(str_detect(trait, 'dependent habitats')) %>%
  mutate(trait_value = habs,
         trait_value = str_split(habs, pattern = ';')) %>%
  unnest(trait_value) %>%
  distinct()
  

traitvals_df_valid <- traitvals_df_prehabs %>%
  filter(!str_detect(trait, 'dependent habitats')) %>%
  bind_rows(traitvals_hab) %>%
  rename(trait_value = trait_value)

write_csv(traitvals_df_valid, '_data/trait_vals_valid.csv')
```


## Read in Excel sheets one at a time

### Original traits data

convert to long format, bind into a single dataframe with minimal processing.

First, check headers to make sure there's not taxonomic information in multiple rows.  The first row of each sheet should be blank, as most sheets have a space before the first category ("Movement").  Even without this space, "Movement" should not have any values of its own, so still blank.

Note that with the additional fish data we will process these separately.

```{r}
check_headers <- function(f, s) {
  ### s <- sheetnames[5]; f <- raw_file
  df <- readxl::read_excel(path = f, sheet = s, n_max = 1) %>%
    gather(hdg, val, -c(1:3)) %>%
    select(hdg, val) %>%
    mutate(sheet = s,
           hdg = as.character(hdg),
           val = as.character(val))
  return(df)
}

raw_file <- here('_raw_data/xlsx/species_traits_all.xlsx')

sheetnames <- readxl::excel_sheets(raw_file)

hdr_check <- lapply(sheetnames, FUN = function(s) check_headers(raw_file, s)) %>%
  bind_rows() %>%
  filter(!is.na(val))

extras <- hdr_check$sheet %>% unique()
```

```{r}
raw_file_turtle <- here('_raw_data/xlsx/chelonia_mydas_updated_traits.xlsx')

sheetnames_turtle <- readxl::excel_sheets(raw_file_turtle)

hdr_check_turtle <- lapply(sheetnames_turtle, FUN = function(s) check_headers(raw_file_turtle, s)) %>%
  bind_rows() %>%
  filter(!is.na(val))

extras_turtle <- hdr_check_turtle$sheet %>% unique()

```

``` {r}
process_sheet <- function(f, s) {
  ### f <- raw_file; s <- sheetnames[10]
  df_raw <- readxl::read_excel(path = f, sheet = s)
  
  ### read in to get "valid" categories
  categories <- read_csv('_data/trait_vals_valid.csv') %>%
    .$category %>% unique()
    
  df <- df_raw %>%
    gather(spp_gp, trait_value, -c(1:3)) %>%
    janitor::clean_names() %>%
    select(-category_units)
  
  df_cleaned <- df %>%
    mutate_if(is.character, tolower) %>%
    mutate(trait = ifelse(is.na(sub_category), category, sub_category),
           category = ifelse(category == 'photosynthetic', 'specialisation', category),
             ### fill in below the "physiological tolerance breadths" category
           category = ifelse(!category %in% categories, NA, category)) %>%
    fill(category, .direction = 'down') %>%
    select(category, trait, spp_gp, trait_value) %>%
    filter(!is.na(trait)) %>%
      ### drops rows with no trait listed
    filter(trait != category) %>%
      ### drops category headings in the traits column
    mutate(sheet = s)
  
  return(df_cleaned)
}


raw_data_df <- lapply(sheetnames, 
                      FUN = function(s) {
                        process_sheet(raw_file, s)
                      }) %>%
  bind_rows() %>%
  ### fix typos and minor fixes:
  mutate(trait_value = str_replace_all(trait_value, 'unkown', 'unknown'),
         trait_value = str_replace_all(trait_value, 'demerssal', 'demersal'),
         trait_value = str_replace_all(trait_value, 'esturary', 'estuary'),
         trait_value = str_replace_all(trait_value, 'interntidal', 'intertidal'),
         trait_value = str_replace_all(trait_value, 'lentgh', 'length'),
         trait_value = str_replace_all(trait_value, 'neary sessile', 'nearly sessile'),
         trait_value = str_replace_all(trait_value, 'this sp,', 'this sp.'),
         trait = str_replace_all(trait, '\\(pld\\)$', '(pld) exposure'))

x <- raw_data_df %>%
  select(trait_value) %>% distinct()

```

### New fish traits data

New species traits from Beth in transposed format; new code to incorporate.
```{r fish from beth polidoro}
fish_beth_raw <- readxl::read_excel(here('_raw_data/xlsx/fish_traits_add_bp.xlsx'),
                                sheet = 'srli_fish_traits') %>%
  setNames(tolower(names(.))) %>%
  slice(-c(1, 2)) ### drop empty rows

category_df <- raw_data_df %>%
  select(category, trait) %>%
  distinct()

fish_beth <- fish_beth_raw %>%
  gather(trait, trait_value, -order, -family, -species) %>%
  select(-order, -family, spp_gp = species) %>%
  filter(!is.na(trait_value)) %>%
  mutate(sheet = 'fish',
         spp_gp = tolower(spp_gp),
         trait = str_remove(trait, '\\.\\.\\.[0-9]+')) %>%
  full_join(category_df, by = 'trait')
  
x <- fish_beth %>% filter(is.na(category))
# x$trait %>% unique()
y <- fish_beth %>% filter(is.na(trait))

raw_data_df <- raw_data_df %>%
  bind_rows(fish_beth)
```

New species traits from Mike Kingsford in original format; added instead to `spp_traits_all.xlsx` so no need to re-incorporate.
```{r fish from mike kingsford}
fish_mike_raw <- readxl::read_excel(here('_raw_data/xlsx/fish_traits_add_mk.xlsx'))

new_spp <- names(fish_mike_raw)[-c(1:6)] %>% tolower()
# new_spp[!new_spp %in% raw_data_df$spp_gp]

#  [1] "osterhinchus doederleini"    "cephalopholis cyanostogmata" "lutjanus carponatatus"      
#  [4] "scarus schlegeri"            "ctenoscarus bicolor"         "naso vlamingi"              
#  [7] "acanthurus dusumieri"        "gymnothorax javanicus"       "carapus bermudensi"         
# [10] "phycoduras eques"            "gobiodon histro"             "corphaena hippurus"         
# [13] "gadus mohua"

### these are all typos or unaccepted names that in Mike's spreadsheet that
### have been fixed in the up-to-date species trait spreadsheet.
### No action needed.

```

### Bring in updated chelonia mydas traits for case study and merge original file with chelonia mydas traits and filter to only chelonia mydas.
```{r}
raw_data_df_turtle <- lapply(sheetnames_turtle,
                            FUN = function(s) {
                              process_sheet(raw_file_turtle,s)
                            }) %>% 
  bind_rows() %>%
  ### fix typos and minor fixes:
  mutate(trait_value = str_replace_all(trait_value, 'unkown', 'unknown'),
         trait_value = str_replace_all(trait_value, 'demerssal', 'demersal'),
         trait_value = str_replace_all(trait_value, 'esturary', 'estuary'),
         trait_value = str_replace_all(trait_value, 'interntidal', 'intertidal'),
         trait_value = str_replace_all(trait_value, 'lentgh', 'length'),
         trait_value = str_replace_all(trait_value, 'neary sessile', 'nearly sessile'),
         trait_value = str_replace_all(trait_value, 'this sp,', 'this sp.'),
         trait = str_replace_all(trait, '\\(pld\\)$', '(pld) exposure'))

raw_data_df <- bind_rows(raw_data_df, raw_data_df_turtle) %>%
  filter(grepl("chelonia mydas", spp_gp))

```
## clean up troublesome traits

Check raw data against valid category values.

* Set all values to lower case to avoid case sensitivity.
* replace "n/a" with "na"
* Ditch numeric commas.
* Split trait values on semicolons to separate multiple cell values; then unnest and str_trim.



### zone

```{r clean up zone traits}
traitvals_df_clean <- read_csv('_data/trait_vals_valid.csv') %>%
  mutate(cat_val_clean = clean_trait(trait_value))

zone_vals <- traitvals_df_valid %>%
  filter(trait == 'zone') %>%
  .$trait_value %>% paste0(collapse = '|')

spp_data_zones <- raw_data_df %>%
  filter(trait == 'zone') %>%
  mutate(trait_value = str_replace(trait_value, 'pelagic larvae', ''),
           ### drop sponge pelagic larvae from zones!
         trait_value = str_replace(trait_value, 'benthopelagic', 'pelagic;benthic'),
         trait_value = str_replace(trait_value, 'pelagic', 'neritic;oceanic'),
         trait_value = str_replace(trait_value, 'coastal', 'neritic'),
         trait_value = str_replace(trait_value, 'range across all', zone_vals)) %>%
  ### fix misspelling:
  mutate(tmp = str_extract_all(trait_value, zone_vals)) %>%
  unnest(tmp) %>%
  mutate(trait_val_clean = clean_trait(tmp)) %>%
  clean_up_mess()
```

### adult mobility

``` {r clean up mobility traits}
spp_data_mobility <- raw_data_df %>%
  filter(trait == 'adult mobility') %>%
  mutate(tmp = str_replace_all(tolower(trait_value), '[^a-z ]+', ';')) %>%
  ### split and unnest, trim and clean:
  mutate(tmp = str_split(tmp, ';')) %>%
  unnest(tmp)

#Checking distinct count for each group in mobility
mobility_distinct_check<-spp_data_mobility %>%
  group_by(tmp) %>%
  summarize(count = n()) %>%
  mutate(total = sum(count)) %>%
  arrange(desc(count))

DT::datatable(mobility_distinct_check)
```

``` {r Continuing clean-up for mobility traits}
mob_vals <- traitvals_df_valid %>%
  filter(trait == 'adult mobility') %>%
  .$trait_value %>% paste0(collapse = '|')
# "sessile|nearly sessile/sedentary|passive|vertical migrator|mobile resident|horizontal migrator|nomadic"
  
### clean up some inconsistencies
spp_data_mobility <- spp_data_mobility %>%
  mutate(tmp = str_replace_all(tmp, 'hor[a-z]+', 'horizontal'),
         tmp = str_replace_all(tmp, 'vertic[a-z]+', 'vertical'),
         tmp = str_replace_all(tmp, 'migrat[a-z]+', 'migrator'),
         tmp = str_replace_all(tmp, 'residen[a-z]+', 'resident'),
         tmp = str_replace_all(tmp, 'nearly sessile|sedentary', 'nearly sessile/sedentary')) %>%
  # mutate(tmp = ifelse(str_detect(tmp, mob_vals),
  #                                 str_extract_all(tmp, mob_vals),
  #                                 tmp)) %>%
  mutate(tmp = str_extract_all(tmp, mob_vals)) %>%
  unnest(tmp) %>%
  mutate(trait_val_clean = clean_trait(tmp)) %>%
  clean_up_mess()

spp_data_mobility$trait_val_clean %>% table()
```


``` {r Re-check mobility counts}
mobility_distinct_cleancheck<-spp_data_mobility %>%
  group_by(trait_val_clean) %>%
  summarize(count = n()) %>%
  mutate(total = sum(count)) %>%
  arrange(desc(count))

DT::datatable(mobility_distinct_cleancheck)
```

### depth (min/max)

Add atmosphere zone (exposure to air, for air temp and similar stressors that act out of water) for species who:

* inhabit intertidal zone AND are sessile e.g. anemones, mussels, barnacles, etc. but NOT fish, cephs, etc that can track the tides
* have terrestrial/marine dependent habitats
* have flight

``` {r clean up depth traits}
depth_vals <- traitvals_df_valid %>%
  filter(trait == 'depth (min/max)') %>%
  .$trait_value %>% paste0(collapse = '|')

spp_data_depth_text <- raw_data_df %>%
  filter(trait == 'depth (min/max)') %>%
  mutate(trait_value = str_replace(trait_value, 'range across all', depth_vals),
         tmp = str_replace_all(tolower(trait_value), '[^a-z0-9 ]+', ';'),
         tmp = str_extract_all(tmp, depth_vals)) %>%
  unnest(tmp) %>%
  mutate(trait_val_clean = clean_trait(tmp)) %>%
  clean_up_mess()

### Depth zones:
### Epipelagic (0-200 m)
### Mesopelagic (200 m– 6-700 m)
### Bathypelagic (700 m – 1-2000 m)
### Abyssopelagic (2000 m – 4-6000 m)
### Hadopelagic (6000 m – 10,000 m)
depth_df <- data.frame(d = 0:20000) %>%
  mutate(zone = case_when(d <=  200 ~ 'epipelagic',
                          d <=  700 ~ 'mesopelagic',
                          d <= 2000 ~ 'bathypelagic',
                          d <= 6000 ~ 'abyssopelagic',
                          TRUE      ~ 'hadopelagic')) %>%
  mutate(zone = fct_inorder(zone))
zone_fct <- levels(depth_df$zone)

# spp_data_depth_num <- raw_data_df %>%
#   filter(trait == 'depth (min/max)') %>%
#   filter(str_detect(trait_value, '[0-9]')) %>%
#   rowwise() %>%
#   mutate(nums = str_extract_all(trait_value, '[0-9]+'),
#          min_d = min(as.numeric(unlist(nums))),
#          max_d = max(as.numeric(unlist(nums)))) %>%
#   left_join(depth_df %>% rename(min_z = zone), by = c('min_d' = 'd')) %>%
#   left_join(depth_df %>% rename(max_z = zone), by = c('max_d' = 'd')) %>%
#   mutate(zones = paste0(zone_fct[as.integer(min_z):as.integer(max_z)], collapse = ';'),
#          zones = str_split(zones, ';')) %>%
#   unnest(zones) %>%
#   select(category, trait, trait_value, trait_val_clean = zones, spp_gp, sheet) %>%
#   distinct()
  


### add air category
spp_sessile <- spp_data_mobility %>%
  filter(trait_val_clean == 'sessile') %>%
  select(spp_gp, sheet) %>%
  distinct()
  
spp_intertidal <- spp_data_zones %>%
  filter(spp_gp %in% spp_sessile$spp_gp) %>%
  filter(trait_val_clean == 'intertidal') %>%
  select(spp_gp, sheet) %>%
  distinct()
spp_terr_marine <- raw_data_df %>%
  filter(str_detect(trait, 'terrestrial and marine')) %>%
  filter(str_detect(tolower(trait_value), '^y')) %>%
  select(spp_gp, sheet) %>%
  distinct()
spp_flight <- raw_data_df %>%
  filter(str_detect(trait, 'flight')) %>%
  filter(str_detect(tolower(trait_value), '^y')) %>%
  select(spp_gp, sheet) %>%
  distinct()
spp_atm <- bind_rows(spp_flight, spp_intertidal, spp_terr_marine) %>%
  mutate(spp_gp = str_split(spp_gp, ';')) %>%
  unnest(spp_gp) %>%
  distinct()
  
spp_data_atm <- spp_atm %>%
  mutate(category = 'spatial scale of species',
         trait    = 'depth (min/max)',
         trait_value = '(sessile AND intertidal) OR flight OR marine/terrestrial', 
         trait_val_clean = 'air')

spp_data_depth <- spp_data_depth_text %>%
  #bind_rows(spp_data_depth_num) %>%
  bind_rows(spp_data_atm) %>%
  distinct()

# x <- spp_data_depth %>%
#   anti_join(spp_data_depth1)

spp_data_depth$trait_val_clean %>% table()
```

```{r}
x <- spp_data_depth %>%
  filter(trait_val_clean == 'air') %>%
  group_by(sheet) %>%
  summarize(n_spp_gps = n_distinct(spp_gp))
knitr::kable(x)
```

### photosynthetic

``` {r clean up photosynthetic traits}
spp_data_photo <- raw_data_df %>%
  filter(trait == 'photosynthetic') %>%
  mutate(trait_val_clean = ifelse(str_detect(trait_value, '^y'), 'yes', 'no')) %>%
  clean_up_mess()

spp_data_photo$trait_val_clean %>% table()

```

### thermal sensitivity

#### Ocean warming:

``` {r clean up thermal sensitivity traits}
### ocean warming values are 5°C-10-15-20-25-30°C
spp_data_thermal1 <- raw_data_df %>%
  filter(str_detect(trait, 'thermal - sensitivity to ocean warming')) %>%
  rowwise() %>%
  mutate(trait_num = str_extract_all(trait_value, '[0-9]+')) %>%
  unnest(trait_num) %>%
  group_by(category, trait, trait_value, spp_gp, sheet) %>% 
  summarize(trait_num = max(as.integer(trait_num), na.rm = TRUE)) %>%
  mutate(trait_num = floor(trait_num / 5) * 5,
         trait_num = ifelse(trait_num > 30, 30, trait_num),
         trait_num = ifelse(is.infinite(trait_num), NA, trait_num)) %>%
  mutate(tmp = ifelse(!is.na(trait_num), paste0(trait_num, 'c'),
                      tolower(trait_value))) %>%
  mutate(trait_val_clean = ifelse(str_detect(tmp, 'broad|low sensitivity'), 
                                  '30c', tmp)) %>%
  clean_up_mess()

spp_data_thermal1$trait_val_clean %>% table()

```

#### Heat spikes:

``` {r}
### heat spike values are yes/no
spp_data_thermal2 <- raw_data_df %>%
  filter(str_detect(trait, 'thermal - sensitivity to heat spikes')) %>%
  mutate(tmp = tolower(trait_value),
         tmp = case_when(str_detect(tmp, 'broad') ~ 'no',
                         str_detect(tmp, 'narrow|sensitive') ~ 'yes',
                         str_detect(tmp, 'can recover') ~ 'no',
                         str_detect(tmp, 'not likely') ~ 'no',
                         TRUE ~ tmp)) %>%
  mutate(trait_val_clean = str_replace_all(tmp, '[^a-z0-9]+', '_')) %>%
  clean_up_mess()

spp_data_thermal2$trait_val_clean %>% table()

spp_data_thermal <- bind_rows(spp_data_thermal1, spp_data_thermal2)
```

### pH

``` {r clean up pH traits}
ph_vals <- traitvals_df_valid %>%
  filter(trait == 'ph') %>%
  .$trait_value %>% paste0(collapse = '|')

### ph values are <7.4|7.5-7.7|7.8-8.2
spp_data_ph <- raw_data_df %>%
  filter(trait == 'ph') %>%
  mutate(tmp1 = str_replace_all(trait_value, '\\(.+\\)', ''),
         tmp1 = str_split(tmp1, ';|,')) %>%
  unnest(tmp1) %>%
  ### get mean - find into which category the mean falls
  rowwise() %>%
  mutate(trait_mean = str_extract_all(tmp1, '[0-9.]+') %>% unlist() %>% as.numeric() %>% mean(),
         trait_mean = ifelse(is.infinite(trait_mean), NA, trait_mean)) %>%
  ungroup() %>%
  mutate(tmp = case_when(is.na(trait_mean) ~ tolower(tmp1),
                         trait_mean < 7.5 ~ '<7.4',
                         trait_mean < 7.8 ~ '7.5-7.7',
                         TRUE ~ '7.8-8.2')) %>%
  mutate(tmp = case_when(str_detect(tmp, 'broad|not sensitive') ~ '<7.4', 
                         str_detect(tmp, 'narrow')              ~ '7.8-8.2',
                         str_detect(tmp, 'no|not directly')     ~ '<7.4', ### assume broad
                         TRUE                                   ~ tmp)) %>%
  mutate(trait_val_clean = clean_trait(tmp)) %>%
  clean_up_mess()

spp_data_ph$trait_val_clean %>% table()
```

### biomineral

#### structure location

``` {r clean up biomineral traits}
biomin1_vals <- traitvals_df_valid %>%
  filter(trait == 'calcium carbonate structure location') %>%
  .$trait_value %>% paste0(collapse = '|')

spp_data_biomin1 <- raw_data_df %>%
  filter(trait == 'calcium carbonate structure location') %>%
  rowwise() %>%
  mutate(tmp = tolower(trait_value),
         tmp = case_when(str_detect(tmp, 'ext.+cover') ~ 'external with a cover',
                         str_detect(tmp, 'calcium carbonate') ~ 'external',
                         str_detect(tmp, biomin1_vals) ~ str_extract_all(tmp, biomin1_vals) %>% 
                                  unlist() %>% paste0(collapse = ';'),
                         str_detect(tmp, '^no') ~ 'none',
                         TRUE ~ tmp)) %>%
  mutate(trait_val_clean = clean_trait(tmp)) %>%
  clean_up_mess()

spp_data_biomin1$trait_val_clean %>% table()
```

#### biomineral

```{r}
biomin2_vals <- traitvals_df_valid %>%
  filter(trait == 'biomineral') %>%
  .$trait_value %>% paste0(collapse = '|')

spp_data_biomin2 <- raw_data_df %>%
  filter(trait == 'biomineral') %>%
  rowwise() %>%
  mutate(tmp = tolower(trait_value),
         tmp = str_replace(tmp, 'hi mg', 'high mg'),
         tmp = str_replace(tmp, 'chitin(/caco3)?', 'chitin/caco3 mix'),
         tmp = case_when(str_detect(tmp, biomin2_vals) ~ str_extract_all(tmp, biomin2_vals) %>% 
                                  unlist() %>% paste0(collapse = ';'),
                         TRUE ~ NA_character_)) %>%
  mutate(trait_val_clean = clean_trait(tmp)) %>%
  clean_up_mess()

spp_data_biomin2$trait_val_clean %>% table()

spp_data_biomin <- bind_rows(spp_data_biomin1, spp_data_biomin2)
```

### body size

``` {r clean up body size}
bodysize_vals <- traitvals_df_valid %>%
  filter(trait == 'adult body mass/body size') %>%
  .$trait_value %>% paste0(collapse = '|')

spp_data_bodysize <- raw_data_df %>%
  filter(trait == 'adult body mass/body size') %>%
  mutate(tmp = str_replace_all(trait_value, '.+:', ''),
         tmp = clean_trait(tmp),
         tmp = str_replace(tmp, ' +\\(.+\\)', ''),
         tmp = ifelse(str_detect(tmp, 'not large'), '0.5-49mm;50mm-999mm', tmp), ### demospongiae
         tmp = ifelse(str_detect(tmp, 'large'), '>1000mm', tmp),
         tmp = str_replace(tmp, '>?1000[^a-z]|>?1000$', '>1000mm'),
         tmp = str_replace_all(tmp, '(?<=[0-9])$', 'mm'),
         tmp = str_replace_all(tmp, '(?<=[0-9])(?=[^a-z0-9\\.])', 'mm'),
         tmp = str_replace_all(tmp, '5.+->', '50mm-999mm;>')) %>%
  rowwise() %>%
  mutate(tmp = ifelse(str_detect(tmp, bodysize_vals), 
                      str_extract_all(tmp, bodysize_vals) %>% unlist() %>% paste(collapse = ';'),
                      NA)) %>%
  mutate(trait_val_clean = clean_trait(tmp)) %>%
  clean_up_mess()

spp_data_bodysize$trait_val_clean %>% table()

```

### fecundity

``` {r clean up fecundity}
### Here, extract numerics from the string, use the min and max values to
### identify the span, then assign categories within that span. UGH

fecund_lims <- traitvals_df_valid %>%
  filter(trait == 'fecundity' & trait_value != 'unknown') %>%
  rowwise() %>%
  mutate(vals = str_extract_all(trait_value, '[0-9]+'),
         fmin = as.integer(vals[1]), 
         fmax = as.integer(vals[2])) %>%
  ungroup() %>%
  mutate(fmin = ifelse(trait_value == '<1', 0, fmin),
         fmax = ifelse(trait_value == '<1', 1, fmax),
         fmax = ifelse(trait_value == '>10000', 10001, fmax)) %>%
  select(trait_value, fmin, fmax) %>%
  arrange(fmin) %>%
  mutate(gp = 1:n(),
         ftmp = NA) %>% ### tmp
  group_by(gp) %>%
  complete(ftmp = fmin:(fmax-1), nesting(trait_value, fmin, fmax)) %>%
  filter(!is.na(ftmp)) %>%
  ungroup() %>%
  mutate(fmin = ftmp,
         fmax = ftmp+1,
         fmax = ifelse(fmax > 10000, Inf, fmax)) %>%
  distinct()
  

spp_data_fecund <- raw_data_df %>%
  filter(trait == 'fecundity') %>%
  filter(!is.na(trait_value)) %>%
  rowwise() %>%
  mutate(tmp = clean_trait(trait_value),
         tmp = str_replace_all(tmp, 'to ', '-'),
         tmp = str_split(tmp, ';')) %>%
  unnest(tmp) %>%
  group_by(spp_gp, sheet) %>%
  mutate(tmp_sep = 1:n()) %>%
  rowwise() %>%
  mutate(vals = str_extract_all(tmp, '[0-9]+')) %>%
  ungroup() %>%
  unnest(vals) %>%
  group_by(spp_gp, sheet, tmp_sep) %>%
  mutate(vals = as.integer(vals),
         fmin = min(vals),
         fmin = ifelse(str_detect(tmp, '<1[^0-9]|<1$'), 0, fmin),
         fmax = max(vals),
         fmax = ifelse(str_detect(tmp, '>10000'), Inf, fmax)) %>%
  left_join(fecund_lims %>% select(fmin, gp_min = gp), by = 'fmin') %>%
  left_join(fecund_lims %>% select(fmax, gp_max = gp), by = 'fmax') %>%
  mutate(gp_max = ifelse(gp_min > gp_max, gp_min, gp_max),
         gp = NA) %>% ### tmp
  distinct() %>%
  group_by(spp_gp, sheet, tmp_sep, category, trait, trait_value) %>%
  complete(gp = min(gp_min):max(gp_max)) %>%
  ungroup() %>%
  filter(!is.na(gp)) %>%
  left_join(fecund_lims %>% select(gp, trait_val_clean = trait_value), by = 'gp') %>%
  select(category, trait, trait_value, spp_gp, sheet, trait_val_clean) %>%
  distinct()

spp_data_fecund$trait_val_clean %>% table()

```

### post-birth/hatch dependence

``` {r clean up parental dependence}

spp_data_par_dep <- raw_data_df %>%
  filter(str_detect(trait, 'post-birth/hatching.+dependence')) %>%
  mutate(trait = 'post-birth/hatching parental dependence') %>%
  filter(!is.na(trait_value)) %>%
  rowwise() %>%
  mutate(tmp = clean_trait(trait_value),
         tmp = str_replace_all(tmp, 'to ', '-'),
         tmp = ifelse(str_detect(tmp, 'hatchling maternal'), 'month-year', tmp),
         tmp = ifelse(str_detect(tmp, '8-17months'), 'month-year;>year', tmp),
         tmp = ifelse(str_detect(tmp, 'month-year-exception'), 'month-year;>year', tmp),
         tmp = ifelse(str_detect(tmp, 'some nursing'), 'week-month', tmp),
         tmp = ifelse(tmp == 'no' | tmp == 'nil', 'na', tmp),
         tmp = ifelse(tmp == 'week -two months', 'week-month', tmp),
         tmp = ifelse(tmp == 'yes-live birth', 'na', tmp),
         tmp = str_split(tmp, ';')) %>%
  unnest(tmp) %>%
  select(category, trait, trait_value, spp_gp, sheet, trait_val_clean = tmp) %>%
  distinct()

spp_data_par_dep$trait_val_clean %>% table()

```

### global pop size

``` {r clean up global pop size}
### Here, extract numerics from the string, use the min and max values to
### identify the span, then assign categories within that span. UGH

pop_lims <- traitvals_df_valid %>%
  filter(trait == 'global population size' & trait_value != 'unknown') %>%
  mutate(trait_value = str_replace_all(trait_value, '(?<=[0-9])k', '000'),
         trait_value = str_replace_all(trait_value, '(?<=[0-9])m', '000000')) %>%
  rowwise() %>%
  mutate(vals = str_extract_all(trait_value, '[0-9]+'),
         fmin = as.integer(vals[1]), 
         fmax = as.integer(vals[2])) %>%
  ungroup() %>%
  mutate(fmin = ifelse(trait_value == '<1000', 0, fmin),
         fmax = ifelse(trait_value == '<1000', 1000, fmax),
         fmax = ifelse(trait_value == '>1000000', 1000001, fmax)) %>%
  select(trait_value, fmin, fmax) %>%
  arrange(fmin) %>%
  mutate(gp = 1:n(),
         ftmp = NA) %>% ### tmp
  group_by(gp) %>%
  complete(ftmp = fmin:(fmax-1), nesting(trait_value, fmin, fmax)) %>%
  filter(!is.na(ftmp)) %>%
  ungroup() %>%
  mutate(fmin = ftmp,
         fmax = ftmp+1,
         fmax = ifelse(fmax > 1000000, Inf, fmax)) %>%
  distinct()
  

spp_data_pop1 <- raw_data_df %>%
  filter(trait == 'global population size') %>%
  filter(!is.na(trait_value)) %>%
  rowwise() %>%
  mutate(tmp = str_split(clean_trait(trait_value), ';')) %>%
  unnest(tmp) %>%
  mutate(tmp = str_replace_all(tmp, '(?<=[0-9])k', '000'),
         tmp = str_replace_all(tmp, '(?<=[0-9])m', '000000'),
         tmp = str_replace_all(tmp, 'unsure|data deficient', 'unknown'))

spp_data_pop2 <- spp_data_pop1 %>%
  filter(str_detect(tmp, 'unknown')) %>%
  mutate(trait_val_clean = 'unknown') %>%
  clean_up_mess()

spp_data_pop3 <- spp_data_pop1 %>%
  filter(!str_detect(tmp, 'unknown')) %>%
  group_by(spp_gp, sheet) %>%
  mutate(tmp_sep = 1:n()) %>%
  rowwise() %>%
  mutate(vals = str_extract_all(tmp, '[0-9]+')) %>%
  ungroup() %>%
  unnest(vals) %>%
  group_by(spp_gp, sheet, tmp_sep) %>%
  mutate(vals = as.integer(vals),
         fmin = min(vals),
         fmin = ifelse(str_detect(tmp, '<1[^0-9]|<1$'), 0, fmin),
         fmax = max(vals),
         fmax = ifelse(str_detect(tmp, '>10000'), Inf, fmax)) %>%
  left_join(pop_lims %>% select(fmin, gp_min = gp), by = 'fmin') %>%
  left_join(pop_lims %>% select(fmax, gp_max = gp), by = 'fmax') %>%
  mutate(gp_max = ifelse(gp_min > gp_max, gp_min, gp_max),
         gp = NA) %>% ### tmp
  distinct() %>%
  group_by(spp_gp, sheet, tmp_sep, category, trait, trait_value) %>%
  complete(gp = min(gp_min):max(gp_max)) %>%
  ungroup() %>%
  filter(!is.na(gp)) %>%
  left_join(pop_lims %>% select(gp, trait_val_clean = trait_value) %>% distinct(), by = 'gp') %>%
  select(category, trait, trait_value, spp_gp, sheet, trait_val_clean) %>%
  distinct() %>%
  ### convert back to m and k versions... UGH
  mutate(trait_val_clean = str_replace_all(trait_val_clean, '000000', 'm'),
         trait_val_clean = str_replace_all(trait_val_clean, '000(?=[^0-9])|000$', 'k'),
         trait_val_clean = str_replace_all(trait_val_clean, '<1k', '<1000'))

spp_data_pop <- bind_rows(spp_data_pop2, spp_data_pop3)

spp_data_pop$trait_val_clean %>% table()

```

### max age

``` {r clean up max age}
### Here, extract numerics from the string, use the min and max values to
### identify the span, then assign categories within that span. UGH

age_lims <- traitvals_df_valid %>%
  filter(trait == 'max age' & trait_value != 'unknown') %>%
  rowwise() %>%
  mutate(vals = str_extract_all(trait_value, '[0-9]+'),
         fmin = as.integer(vals[1]), 
         fmax = as.integer(vals[2])) %>%
  ungroup() %>%
  mutate(fmin = ifelse(str_detect(trait_value, '3month'), 0, fmin),
         fmax = ifelse(str_detect(trait_value, '3month'), 1, fmax),
         fmax = ifelse(trait_value == '>100yrs', 101, fmax)) %>%
  select(trait_value, fmin, fmax) %>%
  arrange(fmin, trait_value) %>%
  mutate(gp = 1:n(),
         ftmp = NA) %>% ### tmp
  group_by(gp) %>%
  complete(ftmp = fmin:(fmax-1), nesting(trait_value, fmin, fmax)) %>%
  filter(!is.na(ftmp)) %>%
  ungroup() %>%
  mutate(fmin = ftmp,   fmin = ifelse(trait_value == '3months-1yr', .25, fmin),
         fmax = ftmp+1, fmax = ifelse(trait_value == '<3months', .25, fmax),
         fmax = ifelse(fmax > 100, Inf, fmax)) %>%
  distinct()
  

spp_data_age <- raw_data_df %>%
  filter(trait == 'max age') %>%
  filter(!is.na(trait_value)) %>%
  rowwise() %>%
  mutate(tmp = str_replace_all(trait_value, '.+:|\\(.+\\)', ''),
         tmp = clean_trait(tmp),
         tmp = str_replace_all(tmp, '3month', '0.25'),
         tmp = str_split(tmp, ';')) %>%
  unnest(tmp) 

spp_data_age1 <- spp_data_age %>%
  filter(str_detect(tmp, 'unknown|little known')) %>%
  mutate(trait_val_clean = 'unknown') %>%
  clean_up_mess()

spp_data_age2 <- spp_data_age %>%
  filter(!str_detect(tmp, 'unknown|little known')) %>%
  group_by(spp_gp, sheet) %>%
  mutate(tmp_sep = 1:n()) %>%
  rowwise() %>%
  mutate(vals = str_extract_all(tmp, '[0-9\\.]+')) %>%
  ungroup() %>%
  unnest(vals) %>%
  group_by(spp_gp, sheet, tmp_sep) %>%
  mutate(vals = as.numeric(vals),
         fmin = min(vals),
         fmax = max(vals),
         fmin = ifelse(str_detect(tmp, '<0.25'), 0, fmin),
         fmax = ifelse(fmax > 100, Inf, fmax)) %>%
  left_join(age_lims %>% select(fmin, gp_min = gp), by = 'fmin') %>%
  left_join(age_lims %>% select(fmax, gp_max = gp), by = 'fmax') %>%
  mutate(gp_max = ifelse(gp_min > gp_max, gp_min, gp_max),
         gp = NA) %>% ### tmp
  distinct() %>%
  group_by(spp_gp, sheet, tmp_sep, category, trait, trait_value) %>%
  complete(gp = min(gp_min):max(gp_max)) %>%
  ungroup() %>%
  filter(!is.na(gp)) %>%
  left_join(age_lims %>% select(gp, trait_val_clean = trait_value) %>% distinct(), by = 'gp') %>%
  select(category, trait, trait_value, spp_gp, sheet, trait_val_clean) %>%
  distinct()

spp_data_age <- bind_rows(spp_data_age1, spp_data_age2)
spp_data_age$trait_val_clean %>% table()

```

### reproductive strategy

``` {r clean up reprod strategy}

reprod_strategy_vals <- traitvals_df_valid %>%
  filter(trait == 'reproductive strategy') %>%
  .$trait_value %>% paste0(collapse = '|')
# "sexual dioecious|sexual hermaphrodite|asexual|colonial|unknown"

spp_data_reprod_strategy <- raw_data_df %>%
  filter(trait == 'reproductive strategy') %>%
  filter(!is.na(trait_value)) %>%
  rowwise() %>%
  mutate(tmp = str_replace_all(trait_value, '.+:', ''),
         tmp = clean_trait(tmp),
         tmp = str_split(tmp, ';| and | or |/')) %>%
  unnest(tmp) %>%
  mutate(tmp = case_when(str_detect(tmp, 'dioec') ~ 'sexual dioecious',
                         str_detect(tmp, 'herma?ph') ~ 'sexual hermaphrodite',
                         str_detect(tmp, 'asexual') ~ 'asexual',
                         str_detect(tmp, 'sexual(?!=( dio| herm))') ~ 'sexual dioecious;sexual hermaphrodite',
                         str_detect(tmp, 'heteromorphic') ~ 'asexual;sexual dioecious;sexual hermaphrodite',
                         TRUE ~ tmp)) %>%
  mutate(trait_val_clean = str_split(tmp, ';')) %>%
  unnest(trait_val_clean) %>%
  clean_up_mess()


spp_data_reprod_strategy$trait_val_clean %>% table()

```

``` {r Re-check mobility counts}
reprod_strat_cleancheck<- spp_data_reprod_strategy %>%
  group_by(trait_value) %>%
  summarize(count = n()) %>%
  mutate(total = sum(count)) %>%
  arrange(desc(count))

DT::datatable(reprod_strat_cleancheck)
```

### age to first reproduction/generation time

``` {r clean up gen time}
### Here, extract numerics from the string, use the min and max values to
### identify the span, then assign categories within that span. UGH

gen_time_lims <- traitvals_df_valid %>%
  filter(trait == 'age to 1st reproduction/generation time' & trait_value != 'unknown') %>%
  rowwise() %>%
  mutate(vals = str_extract_all(trait_value, '[0-9]+'),
         fmin = as.integer(vals[1]), 
         fmax = as.integer(vals[2])) %>%
  ungroup() %>%
  mutate(fmin = ifelse(trait_value == '<1yr', 0, fmin),
         fmax = ifelse(trait_value == '<1yr', 1, fmax),
         fmax = ifelse(trait_value == '>20yrs', 21, fmax)) %>%
  select(trait_value, fmin, fmax) %>%
  arrange(fmin) %>%
  mutate(gp = 1:n(),
         ftmp = NA) %>% ### tmp
  group_by(gp) %>%
  complete(ftmp = fmin:(fmax-1), nesting(trait_value, fmin, fmax)) %>%
  filter(!is.na(ftmp)) %>%
  ungroup() %>%
  mutate(fmin = ftmp,
         fmax = ftmp+1,
         fmax = ifelse(fmax > 20, Inf, fmax)) %>%
  distinct()
  

spp_data_gen_time <- raw_data_df %>%
  filter(trait == 'age to 1st reproduction/generation time') %>%
  filter(!is.na(trait_value)) %>%
  rowwise() %>%
  mutate(tmp = str_replace_all(trait_value, '.+:', ''),
         tmp = clean_trait(tmp),
         tmp = str_split(tmp, ';')) %>%
  unnest(tmp) 

spp_data_gen_time1 <- spp_data_gen_time %>%
  filter(str_detect(tmp, 'unknown')) %>%
  mutate(trait_val_clean = 'unknown') %>%
  clean_up_mess()

spp_data_gen_time2 <- spp_data_gen_time %>%
  filter(!str_detect(tmp, 'unknown')) %>%
  group_by(spp_gp, sheet) %>%
  mutate(tmp_sep = 1:n()) %>%
  rowwise() %>%
  mutate(vals = str_extract_all(tmp, '[0-9]+')) %>%
  ungroup() %>%
  unnest(vals) %>%
  group_by(spp_gp, sheet, tmp_sep) %>%
  mutate(vals = as.integer(vals),
         fmin = min(vals),
         fmin = ifelse(str_detect(tmp, '<1[^0-9]|<1$'), 0, fmin),
         fmax = max(vals),
         fmax = ifelse(fmax > 20, Inf, fmax)) %>%
  left_join(gen_time_lims %>% select(fmin, gp_min = gp), by = 'fmin') %>%
  left_join(gen_time_lims %>% select(fmax, gp_max = gp), by = 'fmax') %>%
  mutate(gp_max = ifelse(gp_min > gp_max, gp_min, gp_max),
         gp = NA) %>% ### tmp
  distinct() %>%
  group_by(spp_gp, sheet, tmp_sep, category, trait, trait_value) %>%
  complete(gp = min(gp_min):max(gp_max)) %>%
  ungroup() %>%
  filter(!is.na(gp)) %>%
  left_join(gen_time_lims %>% select(gp, trait_val_clean = trait_value) %>% distinct(), by = 'gp') %>%
  select(category, trait, trait_value, spp_gp, sheet, trait_val_clean) %>%
  distinct()

spp_data_gen_time <- bind_rows(spp_data_gen_time1, spp_data_gen_time2)

spp_data_gen_time$trait_val_clean %>% table()

y <- spp_data_gen_time %>%
  filter(str_detect(trait_value, 'meter|feet|foot'))
```

### lifetime reproductive opportunities

``` {r clean up lifetime reprod opps}
### Here, extract numerics from the string, use the min and max values to
### identify the span, then assign categories within that span. UGH

reprod_lims <- traitvals_df_valid %>%
  filter(trait == 'lifetime # reproductive opportunities' & trait_value != 'unknown') %>%
  rowwise() %>%
  mutate(vals = str_extract_all(trait_value, '[0-9]+'),
         fmin = as.integer(vals[1]), 
         fmax = as.integer(vals[2])) %>%
  ungroup() %>%
  mutate(fmin = ifelse(trait_value == '1', 0, fmin),
         fmax = ifelse(trait_value == '1', 1, fmax),
         fmax = ifelse(trait_value == '>100', 100, fmax)) %>%
  select(trait_value, fmin, fmax) %>%
  arrange(fmin) %>%
  mutate(gp = 1:n(),
         ftmp = NA) %>% ### tmp
  group_by(gp) %>%
  complete(ftmp = fmin:fmax, nesting(trait_value, fmin, fmax)) %>%
  filter(!is.na(ftmp)) %>%
  ungroup() %>%
  mutate(fmin = ftmp,
         fmax = ftmp+1,
         fmax = ifelse(fmax > 100, Inf, fmax)) %>%
  distinct() %>%
  filter(!(trait_value == '51-100' & fmin == 100))
  

spp_data_reprod_opps <- raw_data_df %>%
  filter(trait == 'lifetime # reproductive opportunities') %>%
  filter(!is.na(trait_value)) %>%
  rowwise() %>%
  mutate(tmp = clean_trait(trait_value),
         tmp = str_replace_all(tmp, '\\(.+\\)', ''),
         tmp = str_split(tmp, ';')) %>%
  unnest(tmp) 

spp_data_reprod1 <- spp_data_reprod_opps %>%
  filter(str_detect(tmp, 'unknown')) %>%
  mutate(trait_val_clean = 'unknown') %>%
  clean_up_mess()

spp_data_reprod2 <- spp_data_reprod_opps %>%
  filter(!str_detect(tmp, 'unknown')) %>%
  group_by(spp_gp, sheet) %>%
  mutate(tmp_sep = 1:n()) %>%
  rowwise() %>%
  mutate(vals = str_extract_all(tmp, '[0-9]+')) %>%
  ungroup() %>%
  unnest(vals) %>%
  group_by(spp_gp, sheet, tmp_sep) %>%
  mutate(vals = as.integer(vals),
         fmin = min(vals),
         fmax = max(vals),
         fmin = ifelse(tmp == '1', 0, fmin),
         fmax = ifelse(str_detect(tmp, '>100'), Inf, fmax)) %>%
  left_join(reprod_lims %>% select(fmin, gp_min = gp), by = 'fmin') %>%
  left_join(reprod_lims %>% select(fmax, gp_max = gp), by = 'fmax') %>%
  mutate(gp_max = ifelse(gp_min > gp_max, gp_min, gp_max),
         gp = NA) %>% ### tmp
  distinct() %>%
  group_by(spp_gp, sheet, tmp_sep, category, trait, trait_value) %>%
  complete(gp = min(gp_min):max(gp_max)) %>%
  ungroup() %>%
  filter(!is.na(gp)) %>%
  left_join(reprod_lims %>% select(gp, trait_val_clean = trait_value) %>% distinct(), by = 'gp') %>%
  select(category, trait, trait_value, spp_gp, sheet, trait_val_clean) %>%
  distinct()

spp_data_reprod_opps <- bind_rows(spp_data_reprod1, spp_data_reprod2)

spp_data_reprod_opps$trait_val_clean %>% table()
```

### number of sites

``` {r clean up number of sites}
### Here, sub in reasonable values.  If aggregates but uncertain, assign
### all aggregation values and let the code find the mean value. UGH

nsite_vals <- 'one|few|many|does not aggregate'
### one, few, many, does not aggregate

spp_data_nsite <- raw_data_df %>%
  filter(str_detect(trait, 'number of sites')) %>%
  mutate(tmp = str_replace_all(trait_value, '\\(.+\\)', ''),
         tmp = clean_trait(tmp)) %>%
  rowwise() %>%
  mutate(tmp1 = case_when(str_detect(tmp, 'entire distribution|small') ~ 'one',
                          str_detect(tmp, '^n') ~ 'does not aggregate',
                          str_detect(tmp, nsite_vals) ~ str_extract_all(tmp, nsite_vals) %>% 
                            unlist() %>% paste0(collapse = ';'), ### extract vals in a sec
                          str_detect(tmp, '(?<!not )aggreg|yes') ~ 'one;few;many',
                          TRUE ~ NA_character_),
         tmp1 = str_split(tmp1, ';')) %>%
  ungroup() %>%
  unnest(tmp1) %>%
  mutate(trait_val_clean = clean_trait(tmp1)) %>%
  clean_up_mess()

spp_data_nsite$trait_val_clean %>% table()

```

### extent of occurrence

``` {r clean up eoo}
eoo_vals <- traitvals_df_valid %>%
  filter(trait == 'eoo (range)') %>%
  .$trait_value %>% paste0(collapse = '|')
### [1] "<99"        "100-4999"   "5000-19999" ">20000"     "unknown"   

spp_data_eoo <- raw_data_df %>%
  filter(trait == 'eoo (range)') %>%
  mutate(tmp = str_replace_all(trait_value, ' to ', '-'),
         tmp = str_replace_all(tmp, '20,000\\?', '>20000'),
         tmp = clean_trait(tmp),
         tmp = str_replace_all(tmp, 'km2', ''),
         tmp = str_replace_all(tmp, '38n-72n', '>20000'),
         tmp = str_replace(tmp, '<100', '<99'),
         tmp = str_replace(tmp, '-5000', '-4999'),
         tmp = str_replace(tmp, '-20000', '-19999')) %>%
  mutate(tmp = str_split(tmp, ';')) %>%
  unnest(tmp) %>%
  mutate(trait_val_clean = str_extract_all(tmp, eoo_vals)) %>%
  unnest(trait_val_clean) %>%
  clean_up_mess()



  # mutate(trait_val_clean = clean_trait(tmp1)) %>%
  # clean_up_mess()

spp_data_eoo$trait_val_clean %>% table()

```

### pressure sensitive structures

``` {r clean up pressure sensitive structures}
### Lungs don't seem to be particularly sensitive to high power sonar;
### guessing gas bladders as well.  Assign low sensitivity.

pressure_vals <- traitvals_df_valid %>%
  filter(trait == 'extreme pressure wave sensitive structures') %>%
  .$trait_value
  
spp_data_prs <- raw_data_df %>%
  filter(trait == 'extreme pressure wave sensitive structures') %>%
  filter(!is.na(trait_value)) %>%
  mutate(tmp = str_replace_all(trait_value, 'sensiti?vity', ''),
         tmp = clean_trait(tmp),
         tmp = case_when(tmp == 'probably'      ~ 'medium',
                         tmp == 'probably not'  ~ 'low',
                         str_detect(tmp, 'yes') ~ 'high',
                         str_detect(tmp, '^no')  ~ 'none',
                         str_detect(tmp, 'lung|pneumatocyst') ~ 'low',
                         TRUE ~ tmp)) %>%
  mutate(trait_val_clean = clean_trait(tmp)) %>%
  clean_up_mess()

spp_data_prs$trait_val_clean %>% table()
  
```

### parental investment

``` {r clean up parental investment}
### create a super clean version to match str_extracted values
parent_vals <- traitvals_df_valid %>%
  filter(trait == 'parental investment') %>%
  select(trait_val_clean = trait_value) %>%
  mutate(tmp = str_split(trait_val_clean, ';|/')) %>%
  unnest(tmp) %>%
  mutate(tmp = str_replace_all(tmp, '[^a-z]', '')) %>%
  arrange(desc(tmp)) ### to get unattended to match first!
# "live birth/egg care" "spawner" "egg-layer" "egg-layer unattended"
parent_match <- parent_vals$tmp %>% paste0(collapse = '|')

spp_data_parent <- raw_data_df %>%
  filter(trait == 'parental investment') %>%
  filter(!is.na(trait_value)) %>%
  mutate(tmp = clean_trait(trait_value),
         tmp = str_split(tmp, ';|/|and')) %>%
  unnest(tmp) %>%
  mutate(tmp = str_replace_all(tmp, '[^a-z]', ''),
         tmp = str_extract(tmp, parent_match)) %>%
  left_join(parent_vals, by = 'tmp') %>%
  clean_up_mess()
  
spp_data_parent$trait_val_clean %>% table()
```

### dissolved oxygen

``` {r clean up dissolved oxygen}
do_vals <- traitvals_df_valid %>% 
  filter(trait == 'dissolved oxygen') %>%
  .$trait_value %>% paste0(collapse = '|')
# "low tolerance|medium tolerance|high tolerance|air breathers|unknown"

spp_data_do <- raw_data_df %>%
  filter(trait == 'dissolved oxygen') %>%
  filter(!is.na(trait_value)) %>%
  mutate(tmp = clean_trait(trait_value, drop_parens = FALSE),
         tmp = str_replace(tmp, 'air.?breathers?', 'air breathers'),
         tmp = ifelse(str_detect(tmp, 'air breathers'), 'air breathers', tmp), 
           ### override all others if air breather
         tmp = str_replace(tmp, 'narrow', 'low'),
         tmp = str_replace(tmp, 'broad', 'high tolerance'),
         tmp = str_replace(tmp, 'medium$', 'medium tolerance')) %>%
  mutate(trait_val_clean = str_extract_all(tmp, do_vals)) %>%
  clean_up_mess()
  
spp_data_do$trait_val_clean %>% table()
```

### planktonic larval duration

``` {r clean up pld}
pld_vals <- traitvals_df_valid %>% 
  filter(trait == 'planktonic larval duration (pld) exposure') %>%
  .$trait_value %>% paste0(collapse = '|')
# "<1day|<1week|<1month|<4months|4months-1yr|>1yr|not larvae|holoplankton|unknown"

spp_data_pld <- raw_data_df %>%
  filter(trait == 'planktonic larval duration (pld) exposure') %>%
  filter(!is.na(trait_value)) %>%
  mutate(tmp = str_split(trait_value, ';| or |/')) %>%
  unnest(tmp) %>%
  mutate(tmp = clean_trait(tmp),
         tmp = str_replace_all(tmp, 'not? larva?e', 'not larvae'),
         tmp = str_replace_all(tmp, ' to ', '-'),
         tmp = str_replace_all(tmp, 'year', 'yr'),
         tmp = str_replace_all(tmp, 'days-weeks', '<1month')) %>%
  mutate(trait_val_clean = str_extract_all(tmp, pld_vals)) %>%
  unnest(trait_val_clean) %>%
  clean_up_mess()
  
spp_data_pld$trait_val_clean %>% table()
```

### feeding larva

``` {r clean up feeding larvae}
feeding_vals <- traitvals_df_valid %>% 
  filter(trait == 'feeding larva (post-hatching metamorphosis)') %>%
  .$trait_value %>% paste0(collapse = '|')
# "feeding calcifier|feeding non-calcifier|non-feeding calcifier|non-feeding non-calcifier|no larva|na|unknown"

spp_data_feeding <- raw_data_df %>%
  filter(trait == 'feeding larva (post-hatching metamorphosis)') %>%
  filter(!is.na(trait_value)) %>%
  mutate(tmp = str_split(trait_value, ';')) %>%
  unnest(tmp) %>%
  mutate(tmp = clean_trait(tmp),
         tmp = str_replace_all(tmp, 'not? larva?e?', 'no larva')) %>%
  mutate(trait_val_clean = str_extract_all(tmp, feeding_vals)) %>%
  unnest(trait_val_clean) %>%
  clean_up_mess()
  
spp_data_feeding$trait_val_clean %>% table()
```

### Navigation requirements

``` {r clean up nav reqts}
nav_vals <- c('light', 'sound', 'magnetic', 'unknown', 'no(ne)?', 'other') %>%
  paste0(collapse = '|')

spp_data_nav <- raw_data_df %>%
  filter(str_detect(trait, '^navigation')) %>%
  filter(!is.na(trait_value)) %>%
  mutate(tmp = str_split(trait_value, ';| or |/')) %>%
  unnest(tmp) %>%
  mutate(tmp = clean_trait(tmp),
         trait_val_clean = str_extract_all(tmp, nav_vals)) %>%
  unnest(trait_val_clean) %>%
  clean_up_mess()
  
spp_data_nav$trait_val_clean %>% table()

trait_stem <- 'navigation requirements (%s)'
spp_data_nav_separated <- spp_data_nav %>%
  filter(trait_val_clean %in% c('light', 'sound', 'magnetic')) %>%
  mutate(trait = sprintf(trait_stem, trait_val_clean),
         trait_val_clean = 'yes')
  
```

### Respiration structures trait

``` {r clean respiration structures trait}
spp_data_resp_str <- raw_data_df %>%
  filter(str_detect(trait, '^respiration')) %>%
  filter(!is.na(trait_value)) %>%
  ### fix typo and assign "cuticle" to "skin":
  mutate(trait_value = str_replace(trait_value, 'diffusiion', 'diffusion'),
         trait_value = str_replace(trait_value, 'cuticle', 'skin')) %>%
  ### separate multiple values
  mutate(tmp = str_split(trait_value, ';| or |/')) %>%
  unnest(tmp) %>%
  mutate(trait_val_clean = clean_trait(tmp)) %>%
  clean_up_mess()
  
spp_data_resp_str$trait_val_clean %>% table()

```


### Biomass removal trait

``` {r add biomass removal trait}
	
spp_data_biomass_removal <- raw_data_df %>%
  select(spp_gp, sheet) %>%
  distinct() %>%
  mutate(category = 'biomass removal',
         trait = 'biomass removal sensitivity trait',
         trait_value = 'sensitive',
         trait_val_clean = 'sensitive')
  
```

### Sea surface temperature trait

``` {r add sst rise sensitivity trait}
	
spp_data_sst_rise <- raw_data_df %>%
  select(spp_gp, sheet) %>%
  distinct() %>%
  mutate(category = 'physiological tolerance breadths',
         trait = 'sst rise sensitivity trait',
         trait_value = 'sensitive',
         trait_val_clean = 'sensitive')
  
```


### habitats

```{r clean up habitats}
hab_vals_raw <- readxl::read_excel('_raw_data/xlsx/hab_list.xlsx', sheet = 'habitat_list') %>%
  mutate(hab_type = tolower(hab_type))

hab_vals <- hab_vals_raw %>% 
  .$hab_type %>% 
  paste0(collapse = '|')

spp_data_habs <- raw_data_df %>%
  filter(str_detect(trait, 'dependent habitats')) %>%
  filter(!is.na(trait_value)) %>%
  mutate(tmp = str_split(trait_value, ';| or |/|,|-')) %>%
  unnest(tmp) %>%
  filter(!tmp %in% c('', 'no', 'unknown')) %>%
  mutate(tmp = clean_trait(tmp),
         tmp = str_replace_all(tmp, 'soft benthic|soft sediment', 'soft bottom'),
         tmp = ifelse(str_detect(tmp, 'inter?reef'), 'subtidal soft bottom;deep soft bottom', tmp),
           ### assume "interreefal" refers to sediment
         tmp = ifelse(tmp %in% c('sea ice', 'ice edge'), 'sea ice edge', tmp),
         tmp = ifelse(str_detect(tmp, 'estuar'), 'estuary', tmp),
         valid = str_detect(tmp, hab_vals)) %>%
  rowwise() %>%
  mutate(trait_val_clean = str_extract_all(tmp, hab_vals),
         trait_val_clean = ifelse(is_empty(trait_val_clean), tmp, trait_val_clean)) %>%
  unnest(trait_val_clean) %>%
  clean_up_mess()

```

#### valid habitats

```{r}
spp_data_habs %>%
  filter(trait_val_clean %in% hab_vals_raw$hab_type) %>%
  .$trait_val_clean %>%
  table()
```

#### unmatched habitats

```{r}
spp_data_habs %>%
  filter(!trait_val_clean %in% hab_vals_raw$hab_type) %>%
  .$trait_val_clean %>%
  table()
```

## combine into a relatively clean df

``` {r}

cleaned_traits <- c('zone', 'adult mobility', 
                    'fecundity',
                    'respiration structures',
                    'planktonic larval duration (pld) exposure',
                    'feeding larva (post-hatching metamorphosis)',
                    'reproductive strategy',
                    'parental investment',
                    'global population size',
                    'max age',
                    'depth (min/max)',
                    'number of sites',
                    'number of sites, incl. terrestrial wetlands',
                    'eoo (range)',
                    'age to 1st reproduction/generation time',
                    'lifetime # reproductive opportunities',
                    'thermal - sensitivity to ocean warming - max temps tolerated',
                    'thermal - sensitivity to heat spikes/heat waves',
                    'extreme pressure wave sensitive structures',
                    'adult body mass/body size',
                    'ph',
                    'photosynthetic',
                    'dissolved oxygen',
                    'biomineral',
                    'calcium carbonate structure location',
                    'within-stage dependent habitats + condition',
                    'across-stage dependent habitats + condition',
                    'navigation requirements (sound or light, or magnetic)')

spp_data_clean <- raw_data_df %>%
  filter(!trait %in% cleaned_traits) %>%
  filter(!str_detect(trait, 'post-birth/hatching.+dependence')) %>%
  mutate(trait_val_clean = clean_trait(trait_value)) %>%
  clean_up_mess() %>%
  bind_rows(spp_data_zones) %>%
  bind_rows(spp_data_mobility) %>%
  bind_rows(spp_data_depth) %>%
  bind_rows(spp_data_thermal) %>%
  bind_rows(spp_data_ph) %>%
  bind_rows(spp_data_photo) %>%
  bind_rows(spp_data_biomin) %>%
  bind_rows(spp_data_bodysize) %>%
  bind_rows(spp_data_fecund) %>%
  bind_rows(spp_data_par_dep) %>%
  bind_rows(spp_data_pop) %>%
  bind_rows(spp_data_prs) %>%
  bind_rows(spp_data_gen_time) %>%
  bind_rows(spp_data_reprod_opps) %>%
  bind_rows(spp_data_reprod_strategy) %>%
  bind_rows(spp_data_nsite) %>%
  bind_rows(spp_data_eoo) %>%
  bind_rows(spp_data_pld) %>%
  bind_rows(spp_data_feeding) %>%
  bind_rows(spp_data_do) %>%
  bind_rows(spp_data_parent) %>%
  bind_rows(spp_data_age) %>%
  bind_rows(spp_data_nav_separated) %>%
  bind_rows(spp_data_resp_str) %>%
  bind_rows(spp_data_biomass_removal) %>%
  bind_rows(spp_data_sst_rise) %>%
  bind_rows(spp_data_habs) %>%
  ### split and unnest, trim and clean:
  mutate(trait_val_clean = str_split(trait_val_clean, ';')) %>%
  unnest(trait_val_clean) %>%
  filter(!is.na(sheet))

x <- spp_data_clean %>% select(trait, trait_value, trait_val_clean) %>% distinct()

y <- spp_data_clean %>%
  group_by(sheet, trait) %>%
  summarize(n_spp = n_distinct(spp_gp)) %>%
  group_by(sheet) %>%
  complete(trait = unique(spp_data_clean$trait))

z <- spp_data_clean %>%
  group_by(sheet, trait) %>%
  summarize(n_traits = n_distinct(trait)) 
```

## write out clean species group sheets

Here we need to recombine multiple trait vals into one line, separate using a semicolon for easy splitting later.

```{r, eval = TRUE}
spp_data_out <- spp_data_clean %>%
  filter(trait_val_clean != '') %>%
  mutate(spp_gp = tolower(spp_gp),
         ### drop parentheticals with "was" or "incl":
         spp_gp = str_remove_all(spp_gp, '\\((was|incl).+?\\)'),
         spp_gp = str_split(spp_gp, ';')) %>%
  unnest(spp_gp) %>%
  mutate(spp_gp = str_trim(spp_gp)) %>%
  rename(trait_val_orig = trait_value) %>%
  distinct() %>%
  group_by(category, trait, sheet, spp_gp) %>%
  mutate(trait_val_orig = paste0(trait_val_orig, collapse = ';')) %>%
  distinct() %>%
  group_by(category, trait, sheet, spp_gp, trait_val_orig) %>%
  summarize(trait_val_clean = paste0(trait_val_clean, collapse = ';')) %>%
  ungroup()

sheets <- spp_data_out$sheet %>% unique()

unlink(here('_raw_data/spp_trait_csvs/spp_traits_*.csv'))

for(s in sheets) { ### s <- sheets[1]
  spp_gp_data <- spp_data_out %>%
    filter(sheet == s) %>%
    select(-trait_val_orig, -sheet) %>%
    spread(spp_gp, trait_val_clean)
  
  fstem <- here('_raw_data/spp_trait_csvs/spp_traits_%s.csv')
  write_csv(spp_gp_data, sprintf(fstem, s))
}
```
