==> 1_process_spp_trait_data.Rmd <==
---
title: "Process species trait data"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

library(oharac) ### remotes::install_github('oharac/oharac')
library(tidyverse)
library(here)

```

# Summary

Read in Excel spreadsheets of taxonomic traits filled in by taxon experts.  Process into a long/tidy format with checks on validity of category values.  Identify categories that were assigned non-valid values.

# Data

`_raw_data/all_taxa_trait_data.xlsx' is the raw spreadsheet prepared by Nathalie Butt from the various submissions of the taxa-group experts.  This is broken into individual .csvs for manual editing.

# Methods

Read in csvs one at a time, convert to long format, bind into a single dataframe with minimal processing.

``` {r}

spp_csvs <- list.files(here('_raw_data/spp_trait_csvs'), 
                       pattern = 'spp_traits',
                       full.names = TRUE)

taxa <- basename(spp_csvs) %>% str_replace_all('spp_traits_|.csv', '')

process_csv <- function(f) { ### f <- spp_csvs[1]
  x <- read_csv(f, show_col_types = FALSE) %>%
    gather(spp_gp, trait_value, -category, -trait)
}

spp_data_raw <- lapply(spp_csvs, FUN = process_csv) %>%
  setNames(taxa) %>%
  bind_rows(.id = 'taxon') %>%
  mutate(trait_value = str_split(trait_value, pattern = ';')) %>%
  unnest(trait_value) %>%
  separate(spp_gp, c("genus", "species", "stage"), " ") %>%
  unite("spp_gp","genus":"species", sep = " ") %>%
  mutate( stage = ifelse(is.na(stage), "original", stage))

```

Get the valid trait values from the key and bind with the noted traits to compare validity.  Clean up habitats by replacing non-valid instances with "other" to avoid underweighting valid habitats.  For non-mutually-exclusive traits, if a species or group has multiple values noted, calculate a probability of each value.  For example, if fecundity for a genus is scored as "<1, 1-2, 2-5" score it as if a species in that genus has a 33.3% probability of any particular value.

``` {r}

traitvals_valid <- read_csv('_data/trait_vals_valid.csv')

### These traits are not being used
drop_traits <- c('if one/few, size')

### Convert non-matching habitats to "other"
valid_habs <- traitvals_valid %>%
  filter(str_detect(trait, 'habitats')) %>%
  pull(trait_value) %>% unique()
hab_fix_df <- spp_data_raw %>%
  filter(str_detect(trait, 'habitats')) %>%
  select(category, trait, trait_value) %>%
  distinct() %>%
  mutate(val_new = case_when(trait_value %in% valid_habs ~ trait_value,
                             TRUE ~ 'other'))

spp_validcheck <- spp_data_raw %>%
  filter(!trait %in% drop_traits) %>%
  ### fix "other" habitats
  left_join(hab_fix_df, by = c('category', 'trait', 'trait_value')) %>%
  mutate(trait_value = ifelse(!is.na(val_new), val_new, trait_value)) %>%
  select(-val_new) %>%
  distinct() %>%
  ### identify valid trait values to keep
  left_join(traitvals_valid %>% 
              mutate(valid = TRUE), ### add flag to detect unmatched values
            by = c('category', 'trait', 'trait_value')) %>%
  filter(!is.na(spp_gp)) %>%
  filter(trait_value != '' & !is.na(trait_value)) %>%
  group_by(taxon) %>%
  mutate(n_spp_gps = n_distinct(spp_gp)) %>%
  ### keep "other" habitats as well
  mutate(valid = ifelse(str_detect(trait, 'dependent habitats'), TRUE, valid))

spp_invalid <- spp_validcheck %>%
  filter(is.na(valid)) %>%
  filter(trait_value != 'na') %>%
  select(-valid)
write_csv(spp_invalid, 'invalid_spp_trait_values.csv')

non_mut_excl <- c('adult_mobility', 'biomineral', 
                  'calcium_carbonate_structure_location', 
                  'depth_min_max', 'respiration_structures', 
                  'across_stage_dependent_habitats_condition', 
                  'within_stage_dependent_habitats_condition', 
                  'sub_population_dependence_on_particular_sites', 'zone')

spp_valid <- spp_validcheck %>%
  filter(valid) %>%
  select(-valid) %>%
  mutate(category = str_replace_all(category, '[^A-Za-z0-9]+', '_'),
         category = str_replace_all(category, '^_|_$', ''),
         trait    = str_replace_all(trait, '[^A-Za-z0-9]+', '_'),
         trait    = str_replace_all(trait, '^_|_$', '')) %>%
  mutate(spp_gp = str_replace_all(spp_gp, '\\n', ' ')) %>%
  group_by(trait, spp_gp, stage) %>%
  mutate(trait_prob = ifelse(trait %in% non_mut_excl, 1, 1/n_distinct(trait_value))) %>%
  ungroup()
```

Clean up physiological sensitivity to pH for vertebrates, as well as calcium carbonate structure mineral and stages. This traitset is intended to account for shell-building species. Leave in internal structure (otolith).  Also leave in larval non-calcifying feeder stage for fish.  Finally, eliminate "terrestrial and marine life stages" for elasmobranchs.

```{r}
verts <- c('seabirds', 'marine_mammals', 'elasmobranchs', 'fish', 'reptiles')
spp_valid <- spp_valid %>%
  mutate(trait_value = case_when(taxon %in% verts & trait == 'ph' ~ NA_character_,
                                 taxon %in% verts & trait == 'calcium_carbonate_structure_stages' ~ NA_character_,
                                 taxon %in% verts & trait == 'biomineral' ~ NA_character_,
                                 TRUE ~ trait_value)) %>%
  mutate(trait_value = ifelse(taxon == 'elasmobranchs' & trait == 'terrestrial_and_marine_life_stages', 'no', trait_value))
  
```

Include thermal tolerance ranges from AquaMaps envelopes; these are just temporary and will be overwritten at the species level after downfilling.
```{r}
spp_gps_df <- spp_valid %>%
  select(taxon, spp_gp, n_spp_gps, stage) %>%
  distinct()

spp_gp_thermal_tol <- read_csv(here('_raw_data/aquamaps', 
                                    'aquamaps_spp_gp_thermal_range.csv')) %>%
  rename(spp_gp = name) %>%
  inner_join(spp_gps_df, by = 'spp_gp') %>%
  select(-rank) %>%
  mutate(category = 'physiological_tolerance_breadths')

spp_valid_plus_thermal <- spp_valid %>%
  bind_rows(spp_gp_thermal_tol)

### separating out stages from names and adding 'original' to the Butt et al turtle. Finally duplicating the rows and identifying which ones will be gapfilled in the following markdowns.
spp_valid_plus_thermal2 <- spp_valid_plus_thermal %>% 
  bind_rows(spp_valid_plus_thermal) %>%
  mutate (gap = "no") %>%
  group_by(category, trait, trait_value, stage) %>%
  mutate(gap = replace(gap, duplicated(gap), "yes")) %>% 
  ungroup()

# spp_gp_thermal_tol$spp_gp %>% n_distinct()

write_csv(spp_valid_plus_thermal, '_data/spp_traits_valid.csv')
```



There are `r spp_valid$spp_gp %>% n_distinct()` distinct species/species groups accounted for with at least some valid trait codes.

#### Matched traits - count of spp with matched trait values by trait

``` {r}
x <- spp_valid %>% 
              group_by(trait, taxon) %>% 
              summarize(n_matches = n(),
                        n_spp_gps = first(n_spp_gps),
                        pct_match = round(100 * n_matches / n_spp_gps, 1)) %>% 
              arrange(desc(n_matches), taxon)
DT::datatable(x)
```


#### Trouble spots - count of spp with mismatched trait values by trait

``` {r}

DT::datatable(spp_invalid %>% 
                filter(!str_detect(trait, 'habitat')) %>%
                group_by(trait, taxon) %>% 
                summarize(n_mismatches = n(),
                          n_spp_gps = first(n_spp_gps),
                          pct_mismatch = round(100 * n_mismatches / n_spp_gps, 1)) %>% 
                arrange(desc(n_mismatches), taxon))
```
